requirements.txt
altgraph==0.17
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cx-Oracle==8.0.1
decorator==4.4.2
Django==3.0.6
django-cors-headers==3.5.0
django-pyodbc==1.1.3
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
drf-yasg-examples==0.1.1
et-xmlfile==1.0.1
future==0.18.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
networkx==2.4
numpy==1.19.4
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==3.0.5
packaging==20.4
pefile==2019.4.18
pluggy==0.13.1
py==1.9.0
pycel==1.0b22
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pyodbc==4.0.30
pyparsing==2.4.7
pytest==5.3.5
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
